+++ CHANGELOG.md
@@ -1,6 +1,9 @@
+## 1.2.0
+
+- Feat: added `OnboardSDK.getMetroStatus()` and related classes/enums.
 ## 1.1.0
 
-- Feat: added `filterByStopType()` method to `List<Stop>`
+- Feat: added `filterByStopType()` method to `List<Stop>`.
 ## 1.0.0
 
 - Initial version.